refactor(orders): move delete-confirmation and cancel-order-modal into components
Migrates the last orders modals out of the legacy partials/ folder. Refs #131.
