+++ resources/js/pages/orders/components/delete-confirmation.tsx
